docker_server/node_tracker/requirements.txt
pika==1.3.2
requests==2.32.3
cachetools==7.0.5
python-dotenv==1.2.2